refactor: Move the ethers-typed bridge types out of the shared types module
`bridge/ethereum/types.ts` mixed two types that name `ethers` with ten
fee-estimation types that do not. TypeScript loads a declaration file whole, so
the `BridgeDepositFeeEstimation` unions — which every consumer reaches through
`wallet.deposit()` and its siblings — required the optional `ethers` peer to
resolve.

`EthereumWalletConfig` and `EthereumTransactionDetails` move to
`bridge/ethereum/ethers-types.ts`, which is deliberately not re-exported from
`bridge/ethereum`'s barrel. Neither type is integrator-facing: `BridgeOperator`
builds the config from `ConnectedEthereumWallet.toEthWalletConfig()` and passes
it to the bridge classes, and none of those are exported either.

This removes two of the three routes that reached `ethers` from the package
root. The third is `ConnectedEthereumWallet.toEthWalletConfig()`, whose return
type names it in a public signature; that one needs an API decision and is left
alone here.
